back-end/mqtt_processing/requirements.txt
fastapi
pymongo
paho-mqtt
